social_media_app/backend/pyproject.toml
[project]
name = "social-media-backend"
version = "0.1.0"
description = "FastAPI backend for a Twitter-like social app"
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Social Media App", email = "dev@example.com" }
]

dependencies = [
  "fastapi~=0.115.0",
  "uvicorn[standard]~=0.30.0",
  "pydantic~=2.8.0",
  "prometheus-fastapi-instrumentator~=7.0.0",
  "opentelemetry-sdk~=1.27.0",
  "opentelemetry-instrumentation-fastapi~=0.48b0",
  "opentelemetry-exporter-otlp~=1.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest~=8.3.0",
  "pytest-cov~=5.0.0",
  "httpx~=0.27.0",
  "ruff~=0.6.9",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"

